pyproject.toml
[project]
name = "multi-tool-agent"
version = "0.1.0"
description = "Multi-tool LangChain agent with RAG and Text-to-SQL"
requires-python = ">=3.9,<3.14"
readme = "README.md"

dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-community>=0.4.0",
    "langchain-core>=1.0.0",
    "faiss-cpu>=1.8.0",
    "sqlglot>=25.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "tiktoken>=0.7.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
multi-tool-agent = "multi_tool_agent.main:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
